#17 fix: multiple submit
